requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt 
django-filter
drf-yasg                     
django-redis                 
redis
psycopg2-binary
gunicorn
Faker
dj-database-url       